+++ app/src/main/java/com/sthoray/allright/ui/main/view/BrowseFragment.kt
@@ -13,7 +13,7 @@ import com.sthoray.allright.ui.main.adapter.BrowseAdapter
 import com.sthoray.allright.ui.main.view.MainActivity.Companion.CATEGORY_ID_KEY
 import com.sthoray.allright.ui.main.viewmodel.MainViewModel
 import com.sthoray.allright.ui.search.view.SearchActivity
-import com.sthoray.allright.utils.Resource
+import com.sthoray.allright.util.Resource
 import kotlinx.android.synthetic.main.fragment_browse.*
 
 /**
