+++ app/src/main/java/com/sthoray/allright/ui/main/viewmodel/MainViewModel.kt
@@ -7,8 +7,8 @@ import androidx.lifecycle.viewModelScope
 import com.sthoray.allright.data.model.listing.Category
 import com.sthoray.allright.data.model.main.FeatureCategoriesResponse
 import com.sthoray.allright.data.repository.AppRepository
-import com.sthoray.allright.utils.Internet
-import com.sthoray.allright.utils.Resource
+import com.sthoray.allright.util.Internet
+import com.sthoray.allright.util.Resource
 import kotlinx.coroutines.launch
 import retrofit2.Response
 import java.io.IOException
