+++ app/src/main/java/com/sthoray/allright/util/AppIcon.kt
@@ -1,4 +1,4 @@
-package com.sthoray.allright.utils
+package com.sthoray.allright.util
 
 import com.sthoray.allright.R
